The database models have been adjusted, so the beans (along with all the places where the beans are used) have been adjusted to match it (which includes field setting of the beans, reading the fields of the beans, inputting information for the database from jsp pages, and displaying information from the database in jsp pages.
There is a pseudo notification feature for invoices and restock orders. You can now view new, validated, unfinished, and near deadline invoices. You can also view new, completed, unfinished, and near deadline ROs.

The inventory now updates upon creation of the invoice, not upon completion of the invoice. The stock decreases when an invoice is created, and the stock replenishes when an invoice is cancelled.