You can now set the conditions for viewing Products, Invoices, and Restock Orders with the new "Custom View <object>" feature. This decreased the amount of links in the homePage in exchange for more jsp files. As of now, most of the "get" functions run through the getNewServlet, and hopefully it'll soon hold ALL "get" functions.
Also, some weird committing and pushing problem happened, so the changes made in the last commit(s?) are included in this.

Finally, the getRestockOrder.jsp was truncated, so it had to be deleted and re-typed. Unfortunately, it was accidentally deleted before the contents were copied, so it was re-written completely, hence the differences with the old version.